requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
newsapi-python==0.2.6
pip==22.2.2
pyTelegramBotAPI==4.7.0
requests==2.28.1
setuptools==58.1.0
telegraph==2.1.0
urllib3==1.26.12
